restore: write index kv pairs and data kv pairs to separate engine (#132)
* restore: write index kvs and data kvs to seperate engine

* restore: rename  to

* restore: use single engine file to store index

* restore: make index engine limited by table concurrency

* restore: revert some changes

* restore: modify checkpoint proto

* restore: implement checkpoint for index engine file

* tests: add failpoint for CheckpointStatusIndexImported

* address comment

* address comment

* address comment

+++ cmd/tidb-lightning-ctl/main.go
@@ -176,8 +176,8 @@ func checkpointErrorDestroy(ctx context.Context, cfg *config.Config, tableName s
 
 	for _, table := range targetTables {
 		for engineID := 0; engineID < table.EnginesCount; engineID++ {
-			fmt.Fprintln(os.Stderr, "Closing and cleaning up engine:", table.TableName, engineID)
-			closedEngine, err := importer.UnsafeCloseEngine(ctx, table.TableName, engineID)
+			fmt.Fprintln(os.Stderr, "Closing and cleaning up engine:", table.TableName, int32(engineID))
+			closedEngine, err := importer.UnsafeCloseEngine(ctx, table.TableName, int32(engineID))
 			if err != nil {
 				fmt.Fprintln(os.Stderr, "* Encountered error while closing engine:", err)
 				lastErr = err
